POST method for search form submit
