+++ package.json
@@ -43,7 +43,7 @@
       "last 1 safari version"
     ]
   },
-  "proxy": "http://localhost:3001",
+  "proxy": "http://localhost:5000",
   "devDependencies": {
     "autoprefixer": "^9.8.6",
     "postcss": "^7.0.35",
